api/requirements.txt
fastapi==0.136.1
starlette==0.49.3
uvicorn[standard]==0.34.2
asyncpg==0.30.0
pyjwt[crypto]==2.12.1
httpx==0.28.1
pydantic==2.11.3
pydantic-settings==2.9.1
mistune==3.1.3
python-multipart==0.0.26
logfire[fastapi,asyncpg]==4.25.0
sentry-sdk[fastapi]==2.29.1
pyyaml==6.0.2
aioboto3==14.3.0
openpyxl==3.1.5
beautifulsoup4==4.13.4
lxml==6.1.0
opendataloader-pdf==2.3.0
aiosqlite==0.21.0
watchfiles==1.1.0
